// src/app/shared/theme-points/subthemes/subtheme-bonus.enum.ts
export enum SubthemeBonus {
  Thorns = "Thorns",
  BonusDamage = "BonusDamage",
  RecoveryValueBonus = "RecoveryValueBonus",
  ProtectorAura = "ProtectorAura",
  DamageResist = "DamageResist",
  TempHp = "TempHp",
  RiposteAura = "RiposteAura",
  IsolationDamage = "IsolationDamage",
  ActiveDefense = "ActiveDefense",
  CriticalDamageReduction = "CriticalDamageReduction",
  Agile = "Agile",
  Balanced = "Balanced"
}